% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/averageByGroup.R
\name{average}
\alias{average}
\title{calculate Average By Group}
\usage{
average(data, groupBy, calculateAve, na.rm = TRUE)
}
\arguments{
\item{data}{data frame, the data we are working on}

\item{groupBy}{the variable we are using to group by}

\item{calculateAve}{the variable we want to get the mean}

\item{na.rm}{remove NA value or not, default is TRUE}
}
\value{
return the average for each group
}
\description{
Group the data using the parameter "groupBy" first,
then calculate the average of the parameter "calculateAve" for each groups.
The  parameter "calculateAve" should be a numeric column.
The first argument should be a data frame,
otherwise, the function return 'this function need a data frame as the first argument'.
}
\examples{
average(palmerpenguins::penguins,"year",body_mass_g, TRUE)
average(palmerpenguins::penguins,c("year","island"),body_mass_g, TRUE)

\dontrun{
average(palmerpenguins::penguins,"island",species,TRUE)
}


}
