% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_lm_and_plot.R
\name{correlation_lm_and_plot}
\alias{correlation_lm_and_plot}
\title{Linear model fitting and plotting for 2 numerical variable}
\usage{
correlation_lm_and_plot(data, var1, var2)
}
\arguments{
\item{data}{specifies the data set which the variables are found}

\item{var1}{the first variable in the relationship, has to be numerical}

\item{var2}{the second variable in the relationship, has to be numerical}
}
\value{
a list consisting of a tidied up linear model and a plot
the plot is a baseline scatter plot of the 2 variables, a linear line of best fit is present on top of the scatter plot
}
\description{
Linear model fitting and plotting for 2 numerical variable
}
\examples{
correlation_lm_and_plot(datateachr::cancer_sample, datateachr::cancer_sample$ID, datateachr::cancer_sample$concavity_mean)
correlation_lm_and_plot(datateachr::cancer_sample, datateachr::cancer_sample$concave_points_mean, datateachr::cancer_sample$concavity_mean)
}
